finalize

+++ global/application/views/admin/claim_kerja/index.php
@@ -18,7 +18,7 @@
 
     <!-- Main content -->
     <section class="content">
-    <div class="flash-data" data-flashdata="<?= $this->session->flashdata('notif_claim'); ?>"></div>
+    <div class="flash-data" data-flashdata="<?= $this->session->flashdata('notif'); ?>"></div>
 
         <!-- Default box -->
         <div class="card">
@@ -33,10 +33,9 @@
                         </a>
                     </div>
                 </div>
-
             </div>
             <div class="card-body">
-            <table id="example1" class="table table-bordered table-striped">
+            <table id="example1" class="table table-bordered table-striped wrap">
                                   <thead>
                                       <tr>
                                         <th>No.</th>
@@ -63,12 +62,66 @@
                                         <td><?php if ($c->status == 0){
                                                 echo '<i class="fas fa-hourglass-half"></i> Menunggu disetujui';
                                             }elseif ($c->status == 1){
-                                                echo '<i class="fas fa-check"></i>';
+                                                echo '<i class="fas fa-check"></i> Telah disetujui';
                                             }else{echo "Error";};?></td>
                                         <td>
                                             <a href="<?php echo base_url() ?>assets/upload/claim/<?php echo $c->file_rembes;?>"><i class="fas fa-download" title="Download Invoice"></i></a>
-                                            <a href="<?php base_url()?>edit_claim/<?php echo $c->id_rembes;?>"><i class="fas fa-edit" title="Edit"></i></a>
-                                            <a href="<?php base_url()?>delete_claim/<?php echo $c->id_rembes;?>" onclick="return confirm('Batalkan claim?');"><i class="fas fa-trash" title="Delete"></i></a>
+                                            <?php if($c->status == 0){ ?>
+                                                <a href="<?php base_url()?>edit_claim/<?php echo $c->id_rembes;?>"><i class="fas fa-edit" title="Edit"></i></a>
+                                                <a href="<?php base_url()?>delete_claim/<?php echo $c->id_rembes;?>" onclick="return confirm('Batalkan claim?');"><i class="fas fa-trash" title="Delete"></i></a>
+                                            <?php } ?>    
+                                        </td>
+                                      </tr>
+                                      <?php $no++; endforeach?>
+                                  </tbody>
+                              </table>
+               
+            </div>
+            <!-- /.card-body -->
+        </div>
+        <div class="card">
+            <div class="card-header">
+                <div class="row">
+                    <div class="col-md-12">
+                        <h3 class="card-title">Data Pengajuan Claim Kunjungan Kerja</h3>
+                    </div>
+                </div>
+            </div>
+            <div class="card-body">
+            <table id="example2" class="table table-bordered table-striped">
+                                  <thead>
+                                      <tr>
+                                        <th>No.</th>
+                                          <th width="500">Nama Project</th>
+                                          <th>Jumlah Claim</th>
+                                          <th>Tanggal Claim</th>
+                                          <th>Keterangan</th>
+                                          <th>Status</th>
+                                          <th>Pengusul</th>
+                                          <th>Aksi</th>
+                                      </tr>
+                                  </thead>
+                                  <tbody>
+                                      <tr><?php $no=1; 
+                                        foreach ($claim_all as $d) :?>
+                                        <td><?php echo $no?></td>
+                                        <td><?php echo $d->nama_paket;?></td>
+                                        <td><?php echo rupiah($d->nominal);?></td>
+                                        <td><?php echo $d->tanggal_pengajuan;?></td>
+                                        <td><?php echo $d->keterangan;?></td>
+                                        <td><?php if ($d->status == 0){
+                                                echo '<i class="fas fa-hourglass-half"></i> Menunggu disetujui';
+                                            }elseif ($d->status == 1){
+                                                echo '<i class="fas fa-check"></i> Telah disetujui';
+                                            }else{echo "Error";};?></td>
+                                        <td><?php echo $d->nama;?></td>
+                                        <td>
+                                            <a href="<?php echo base_url() ?>assets/upload/claim/<?php echo $d->file_rembes;?>"><i class="fas fa-download" title="Download Invoice"></i></a>
+                                            <?php if ($d->status == 1){?>
+                                                <a href="<?php base_url()?>cancel_claim/<?php echo $d->id_rembes;?>"><i class="fas fa-times" title="Cancel"></i></a>
+                                            <?php } else {?>
+                                                <a href="<?php base_url()?>acc_claim/<?php echo $d->id_rembes;?>"><i class="fas fa-check" title="Approve"></i></a>
+                                                <?php }?>
                                         </td>
                                       </tr>
                                       <?php $no++; endforeach?>
